v3: address coderabbit review comments on test files
- internal/browser: move OS-launching tests to integration build tag
  (browser_integration_test.go); add openCmd injectable var to all
  platform files so unit tests can stub without real side effects;
  TestOpen_Success stubs with os.Args[0] -test.run=^$ (exits 0, cross-platform)
- internal/gosod: move TestExtract_MkdirAllFails to gosod_unix_test.go
  with //go:build unix to avoid Windows CI failures on chmod semantics
- internal/tint: check error return from h.Handle() instead of discarding
  it with _ in TestHandle_NoColor_ProducesNoEscapeCodes and
  TestHandle_Color_ContainsEscapeCodes
