+++ v3/internal/browser/browser_windows.go
@@ -4,8 +4,12 @@ package browser
 
 import "os/exec"
 
+var openCmd = func(target string) *exec.Cmd {
+	return exec.Command("rundll32", "url.dll,FileProtocolHandler", target)
+}
+
 func open(target string) error {
-	cmd := exec.Command("rundll32", "url.dll,FileProtocolHandler", target)
+	cmd := openCmd(target)
 	if err := cmd.Start(); err != nil {
 		return err
 	}
